// dist/node/publications/publication.d.ts
declare class Publication extends TypedEmitter<PublicationEvents> {
    clients: PublicationClientsManager["clients"];
    communityAddress: string;
    shortCommunityAddress: string;
    communityPublicKey?: string;
    communityName?: string;
// ...
    raw: {
        pubsubMessageToPublish?: PublicationFromDecryptedChallengeRequest;
        unsignedPublicationOptions?: CreatePublicationOptions;
    };
// ...
}